LayoutTests/media/modern-media-controls/watchos-media-controls/watchos-media-controls-layout-expected.txt

Testing the WatchOSMediaControls layout.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


With controls at their default state (paused) only the play button is showing.
PASS mediaControls.children is [mediaControls.playButton]

With controls at pending state, only the activity indicator is showing.
PASS mediaControls.children is [mediaControls.activityIndicator]

With controls at invalid state, only the invalid icon is showing.
PASS mediaControls.children is [mediaControls.invalidButton]

PASS successfullyParsed is true

TEST COMPLETE
